requirements.txt
APScheduler==3.5.3
asn1crypto==0.24.0
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.4.1
cycler==0.10.0
future==0.17.1
geographiclib==1.49
geopy==1.17.0
idna==2.7
ipaddress==1.0.22
kiwisolver==1.0.1
matplotlib==3.0.2
more-itertools==4.3.0
mysql-connector-python==8.0.13
numpy==1.15.4
pandas==0.23.4
patsy==0.5.1
pprint==0.1
presto-python-client==0.5.1
protobuf==3.6.1
psycopg2==2.7.6.1
pycparser==2.19
PyMySQL==0.9.2
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.1
requests-kerberos==0.12.0
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
SQLAlchemy==1.2.14
statsmodels==0.9.0
typing==3.6.6
tzlocal==1.5.1
urllib3==1.24.1
winkerberos==0.7.0
xlrd==1.1.0
